feat(backend): workflow domain self-test issue fixed
* feat(backend): self-test issue fixed


* feat(backend): text process optims


* feat(backend): test for 1


See merge request: !254

+++ backend/domain/workflow/internal/nodes/loop/loop.go
@@ -373,7 +373,7 @@ func (l *Loop) getMaxIter(in map[string]any) (int, error) {
 			return 0, errors.New("incoming LoopCount not present in input when loop type is ByIteration")
 		}
 
-		maxIter = iter.(int)
+		maxIter = int(iter.(int64))
 	case Infinite:
 	default:
 		return 0, fmt.Errorf("loop type not supported: %v", l.config.LoopType)
